requirements.txt
finance-datareader==0.9.96
langchain==0.3.19
langchain-community==0.3.18
langchain-core==0.3.58
langchain-huggingface==0.1.2
langgraph==0.3.2
openai==1.65.2
supabase==2.24.0
tavily-python==0.7.13
torch==2.6.0
transformers==4.57.1
yfinance==0.2.66